pyproject.toml
[tool.poetry]
name = "proxiquery"
version = "0.1.0"
description = ""
authors = ["tamuto <tamuto@infodb.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
faiss-cpu = "^1.8.0"
packaging = "^24.1"
sentence-transformers = "^3.0.1"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"
python-dotenv = {extras = ["cli"], version = "^1.0.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
serve = "dotenv run uvicorn server.main:app"
